home-informations/apis/requirements.txt
beautifulsoup4==4.12.3
duckdb==1.0.0
fastapi==0.111.1
gtfs-realtime-bindings==1.0.0
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pydantic-settings==2.5.2
pyarrow==17.0.0
Requests==2.32.3
